Fixed codacy/coding style issues.
